Fix corrupted images
The images were recovered from the map hosted on SourceForge [1].  The
images were stored in triplea-maps/Project using Git LFS [2], which
requires an additional tool to be able to decode.  As these images are
not over 100 MiB in size, it seems useless to require Git LFS to be able
manipulate them.

[1] http://downloads.sourceforge.net/project/tripleamaps/developer%20resources/map%20makers/Map_Making_Tutorial_Map.zip
[2] https://help.github.com/articles/versioning-large-files/